refactor: refactor common constraint func
