+++ src/service/mod.rs
@@ -3,4 +3,5 @@ pub mod error;
 pub mod record;
 pub mod repository;
 pub mod token;
+pub mod utils;
 pub mod zone;
