+++ src/service/record/read.rs
@@ -1,0 +1,73 @@
+use crate::{
+    database::model::record::{Record, RecordType},
+    log_error,
+    service::{error::ServiceError, repository::RepositoryService},
+};
+
+use super::RecordService;
+
+impl RecordService {
+    pub async fn get_records(zone_name: Option<String>) -> Result<Vec<Record>, ServiceError> {
+        match zone_name {
+            Some(name) => {
+                // Check if zone exists and get zone_id
+                let zone = match RepositoryService::get_zone_by_name(&name).await {
+                    Ok(Some(z)) => z,
+                    Ok(None) => {
+                        return Err(ServiceError::BadRequest(format!(
+                            "Zone with name '{}' not found",
+                            name
+                        )));
+                    }
+                    Err(e) => {
+                        log_error!("Failed to fetch zone: {}", e);
+                        return Err(ServiceError::Internal("Failed to fetch zone".to_string()));
+                    }
+                };
+
+                // Fetch records by zone_id
+                match RepositoryService::get_records_by_zone_id(zone.id).await {
+                    Ok(records) => Ok(records),
+                    Err(e) => {
+                        log_error!("Failed to fetch records for zone {}: {}", name, e);
+                        Err(ServiceError::Internal(format!(
+                            "Failed to fetch records for zone {}",
+                            name
+                        )))
+                    }
+                }
+            }
+            None => {
+                // Fetch all records
+                match RepositoryService::get_all_records().await {
+                    Ok(records) => Ok(records),
+                    Err(e) => {
+                        log_error!("Failed to fetch all records: {}", e);
+                        Err(ServiceError::Internal(
+                            "Failed to fetch all records".to_string(),
+                        ))
+                    }
+                }
+            }
+        }
+    }
+
+    pub async fn get_record(name: &str, record_type: &str) -> Result<Record, ServiceError> {
+        // Validate record type
+        let record_type = RecordType::from_str(record_type).map_err(|_| {
+            ServiceError::BadRequest(format!("Invalid record type: {}", record_type))
+        })?;
+
+        match RepositoryService::get_record_by_name_and_type(name, &record_type).await {
+            Ok(Some(record)) => Ok(record),
+            Ok(None) => Err(ServiceError::NotFound(format!(
+                "Record with name '{}' and type '{}' not found",
+                name, record_type
+            ))),
+            Err(e) => {
+                log_error!("Failed to fetch record: {}", e);
+                Err(ServiceError::Internal("Failed to fetch record".to_string()))
+            }
+        }
+    }
+}
