+++ src/service/zone/mod.rs
@@ -1,0 +1,8 @@
+mod create;
+mod delete;
+mod read;
+pub mod snapshot;
+mod update;
+
+#[derive(Clone)]
+pub struct ZoneService;
